Refactored indexer module to flatten it
