+++ server/conda/filesys/init.go
@@ -3,17 +3,20 @@ package filesys
 import (
 	"log"
 
+	"private-conda-repo/config"
 	"private-conda-repo/indexer"
-	_ "private-conda-repo/indexer/docker"
-	_ "private-conda-repo/indexer/shell"
 )
 
 var indMgr indexer.Indexer
 
 func init() {
-	_id, err := indexer.New()
+	conf, err := config.New()
 	if err != nil {
 		log.Fatal(err)
 	}
-	indMgr = _id
+
+	indMgr, err = indexer.New(conf)
+	if err != nil {
+		log.Fatal(err)
+	}
 }
